+++ lib/dsql.dart
@@ -2,7 +2,12 @@ library dsql;
 
 export 'package:postgres/postgres.dart' hide Result;
 
+export 'src/database.dart';
+export 'src/exceptions.dart';
 export 'src/extensions/iterable.dart';
 export 'src/order_by.dart';
+export 'src/page.dart';
+export 'src/params.dart';
+export 'src/repository.dart';
 export 'src/result.dart';
 export 'src/where.dart';