+++ lib/src/where.dart
@@ -17,10 +17,10 @@ class Where {
   const Where.lte(dynamic value) : this._('<=', value);
 
   const Where.startsWith(String value, {bool ignoreCase = true})
-      : this._(ignoreCase ? 'ILIKE' : 'LIKE', '\$value%');
+      : this._(ignoreCase ? 'ILIKE' : 'LIKE', '$value%');
 
   const Where.endsWith(String value, {bool ignoreCase = true})
-      : this._(ignoreCase ? 'ILIKE' : 'LIKE', '%\$value');
+      : this._(ignoreCase ? 'ILIKE' : 'LIKE', '%$value');
 
   const Where.contains(String value, {bool ignoreCase = true})
       : this._(ignoreCase ? 'ILIKE' : 'LIKE', '%$value%');